packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swimburger"
version = "1.0.0"
description = "Personal intro card for Swimburger"
authors = [{name = "Swimburger"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "pyyaml>=6.0",
]
keywords = ["card", "business-card", "cli", "intro"]

[project.scripts]
swimburger = "swimburger:main"

[project.urls]
Homepage = "https://github.com/swimburger/intro"
